Separate librustc_codegen_ssa module
